+++ src/screens/PlayDetail/Vertical/components/Header.tsx
@@ -1,5 +1,5 @@
-import { memo, useRef, useCallback, useMemo } from 'react'
-import { View, StyleSheet, TouchableOpacity, ScrollView } from 'react-native'
+import { memo, useRef, useCallback, useMemo, useState } from 'react'
+import { View, StyleSheet, TouchableOpacity, ScrollView, Modal } from 'react-native'
 import { Icon } from '@/components/common/Icon'
 import { pop, navigations } from '@/navigation'
 import { useTheme } from '@/store/theme/hook'
@@ -14,6 +14,10 @@ import Btn from './Btn'
 import TimeoutExitBtn from './TimeoutExitBtn'
 import Marquee from './Marquee'
 import StatusBar from '@/components/common/StatusBar'
+import { useSetting } from '@/store/setting/hook'
+import { updateSetting } from '@/core/common'
+import SoundEffectControl from '@/components/player/SoundEffectControl'
+import { setNavActiveId } from '@/core/common'
 
 export const HEADER_HEIGHT = scaleSizeH(_HEADER_HEIGHT)
 
@@ -94,11 +98,23 @@ const Title = () => {
 export default memo(() => {
   const popupRef = useRef<SettingPopupType>(null)
   const statusBarHeight = useStatusbarHeight()
+  const setting = useSetting()
+  const soundEffectEnabled = setting['player.soundEffect.enable'] ?? true
+  
   const back = () => {
     void pop(commonState.componentIds[commonState.componentIds.length - 1]?.id!)
   }
   const showSetting = () => {
     popupRef.current?.show()
+  }
+  const toggleSoundEffect = () => {
+    updateSetting({ 'player.soundEffect.enable': !soundEffectEnabled })
+  }
+  const showSoundEffect = () => {
+    setNavActiveId('nav_setting')
+    import('@/screens/Home/Views/Setting/utils').then(({ pushSettingScreen }) => {
+      pushSettingScreen('sound_effect')
+    })
   }
   return (
     <View
@@ -110,6 +126,9 @@ export default memo(() => {
         <Btn icon="chevron-left" onPress={back} />
         <Title />
         <TimeoutExitBtn />
+        {soundEffectEnabled ? (
+          <Btn icon="volume-higt" onPress={showSoundEffect} />
+        ) : null}
         <Btn icon="slider" onPress={showSetting} />
       </View>
       <SettingPopup ref={popupRef} direction="vertical" />